+++ src/userinterface/GameExec.cpp
@@ -3,7 +3,7 @@
 
 GameExec::GameExec(Window* wp, vector<AI*> aiList, int games) :
 		wp(wp) {
-	board = new Board(aiList.size() > PLAYER_LIMIT);
+	board = new Board((int) aiList.size() > PLAYER_LIMIT);
 	simulationLogger = new SimulationLogger(aiList, *board, games);
 }
